requirements.txt
streamlit
google-generativeai
pdfplumber
python-docx
pandas
plotly
